few changes

+++ server/routers/goalRouter.js
@@ -3,7 +3,7 @@ const Goal = require('../models/goalModel');
 
 router.get('/', async (req, res) => {
     try {
-        goals = await Goal.find();
+        goals = await Goal.find({ user: req.user.id }); //and goals added to
         res.json(goals);
     } catch (err) {
         console.error(err);
@@ -15,7 +15,7 @@ router.post('/', async (req, res) => {
     try {
         const goal = await Goal.create({
             text: req.body.text,
-            //userid?
+            user: req.user.id,
         })
 
         res.json(goal);
@@ -32,6 +32,18 @@ router.put('/:id', async (req, res) => {
         if (!goal) 
             return res.status(400).json({errorMessage: "Error: Goal not found."});
         
+        // Check for user
+        if (!req.user) {
+            res.status(401)
+            throw new Error('User not found')
+        }
+    
+        // Make sure the logged in user matches the goal user
+        if (goal.user.toString() !== req.user.id) {
+            res.status(401)
+            throw new Error('User not authorized')
+        }
+
         const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
             new: true,
         });
@@ -49,6 +61,17 @@ router.delete('/:id', async (req, res) => {
         if (!goal) 
             return res.status(400).json({errorMessage: "Error: Goal not found."});
         
+        // Check for user
+        if (!req.user) {
+        res.status(401)
+        throw new Error('User not found')
+        }
+
+        // Make sure the logged in user matches the goal user
+        if (goal.user.toString() !== req.user.id) {
+            res.status(401)
+            throw new Error('User not authorized')
+        }
         await goal.deleteOne();
 
         res.json(goal);
